+++ packages/illustrations/src/__generated__/spotIcon/data/svgJsMap.ts
@@ -9,17 +9,113 @@ import type { SpotIconName } from '../types/SpotIconName';
  */
 
 const svgJsMap = {
+  advancedTradeProduct: {
+    light: () => require('../svgJs/light/advancedTradeProduct-0').content,
+    dark: () => require('../svgJs/dark/advancedTradeProduct-0').content,
+  },
+  assetHubProduct: {
+    light: () => require('../svgJs/light/assetHubProduct-0').content,
+    dark: () => require('../svgJs/dark/assetHubProduct-0').content,
+  },
+  assetManagementProduct: {
+    light: () => require('../svgJs/light/assetManagementProduct-0').content,
+    dark: () => require('../svgJs/dark/assetManagementProduct-0').content,
+  },
+  base: {
+    light: () => require('../svgJs/light/base-0').content,
+    dark: () => require('../svgJs/dark/base-0').content,
+  },
+  borrowProduct: {
+    light: () => require('../svgJs/light/borrowProduct-0').content,
+    dark: () => require('../svgJs/dark/borrowProduct-0').content,
+  },
+  cloudProduct: {
+    light: () => require('../svgJs/light/cloudProduct-0').content,
+    dark: () => require('../svgJs/dark/cloudProduct-0').content,
+  },
+  coinbase: {
+    light: () => require('../svgJs/light/coinbase-0').content,
+    dark: () => require('../svgJs/dark/coinbase-0').content,
+  },
+  coinbaseOneProduct: {
+    light: () => require('../svgJs/light/coinbaseOneProduct-0').content,
+    dark: () => require('../svgJs/dark/coinbaseOneProduct-0').content,
+  },
+  commerceProduct: {
+    light: () => require('../svgJs/light/commerceProduct-0').content,
+    dark: () => require('../svgJs/dark/commerceProduct-0').content,
+  },
+  custodyProduct: {
+    light: () => require('../svgJs/light/custodyProduct-0').content,
+    dark: () => require('../svgJs/dark/custodyProduct-0').content,
+  },
+  dataMarketplace: {
+    light: () => require('../svgJs/light/dataMarketplace-0').content,
+    dark: () => require('../svgJs/dark/dataMarketplace-0').content,
+  },
+  delegateProduct: {
+    light: () => require('../svgJs/light/delegateProduct-0').content,
+    dark: () => require('../svgJs/dark/delegateProduct-0').content,
+  },
+  derivativesProduct: {
+    light: () => require('../svgJs/light/derivativesProduct-0').content,
+    dark: () => require('../svgJs/dark/derivativesProduct-0').content,
+  },
+  exchangeProduct: {
+    light: () => require('../svgJs/light/exchangeProduct-0').content,
+    dark: () => require('../svgJs/dark/exchangeProduct-0').content,
+  },
+  helpCenterProduct: {
+    light: () => require('../svgJs/light/helpCenterProduct-0').content,
+    dark: () => require('../svgJs/dark/helpCenterProduct-0').content,
+  },
+  institutionalProduct: {
+    light: () => require('../svgJs/light/institutionalProduct-0').content,
+    dark: () => require('../svgJs/dark/institutionalProduct-0').content,
+  },
+  internationalExchangeProduct: {
+    light: () => require('../svgJs/light/internationalExchangeProduct-0').content,
+    dark: () => require('../svgJs/dark/internationalExchangeProduct-0').content,
+  },
+  learningRewardsProduct: {
+    light: () => require('../svgJs/light/learningRewardsProduct-0').content,
+    dark: () => require('../svgJs/dark/learningRewardsProduct-0').content,
+  },
+  nftProduct: {
+    light: () => require('../svgJs/light/nftProduct-0').content,
+    dark: () => require('../svgJs/dark/nftProduct-0').content,
+  },
+  nodeProduct: {
+    light: () => require('../svgJs/light/nodeProduct-0').content,
+    dark: () => require('../svgJs/dark/nodeProduct-0').content,
+  },
+  participateProduct: {
+    light: () => require('../svgJs/light/participateProduct-0').content,
+    dark: () => require('../svgJs/dark/participateProduct-0').content,
+  },
+  paySDKProduct: {
+    light: () => require('../svgJs/light/paySDKProduct-0').content,
+    dark: () => require('../svgJs/dark/paySDKProduct-0').content,
+  },
+  primeProduct: {
+    light: () => require('../svgJs/light/primeProduct-0').content,
+    dark: () => require('../svgJs/dark/primeProduct-0').content,
+  },
+  privateClientProduct: {
+    light: () => require('../svgJs/light/privateClientProduct-0').content,
+    dark: () => require('../svgJs/dark/privateClientProduct-0').content,
+  },
   productCoinbaseCard: {
-    light: () => require('../svgJs/light/productCoinbaseCard-0').content,
-    dark: () => require('../svgJs/dark/productCoinbaseCard-0').content,
+    light: () => require('../svgJs/light/productCoinbaseCard-1').content,
+    dark: () => require('../svgJs/dark/productCoinbaseCard-1').content,
   },
   productCompliance: {
     light: () => require('../svgJs/light/productCompliance-0').content,
     dark: () => require('../svgJs/dark/productCompliance-0').content,
   },
   productEarn: {
-    light: () => require('../svgJs/light/productEarn-0').content,
-    dark: () => require('../svgJs/dark/productEarn-0').content,
+    light: () => require('../svgJs/light/productEarn-1').content,
+    dark: () => require('../svgJs/dark/productEarn-1').content,
   },
   productPro: {
     light: () => require('../svgJs/light/productPro-0').content,
@@ -28,6 +124,34 @@ const svgJsMap = {
   productWallet: {
     light: () => require('../svgJs/light/productWallet-0').content,
     dark: () => require('../svgJs/dark/productWallet-0').content,
+  },
+  rewardsProduct: {
+    light: () => require('../svgJs/light/rewardsProduct-0').content,
+    dark: () => require('../svgJs/dark/rewardsProduct-0').content,
+  },
+  rosettaProduct: {
+    light: () => require('../svgJs/light/rosettaProduct-0').content,
+    dark: () => require('../svgJs/dark/rosettaProduct-0').content,
+  },
+  signInProduct: {
+    light: () => require('../svgJs/light/signInProduct-0').content,
+    dark: () => require('../svgJs/dark/signInProduct-0').content,
+  },
+  stakingProduct: {
+    light: () => require('../svgJs/light/stakingProduct-0').content,
+    dark: () => require('../svgJs/dark/stakingProduct-0').content,
+  },
+  venturesProduct: {
+    light: () => require('../svgJs/light/venturesProduct-0').content,
+    dark: () => require('../svgJs/dark/venturesProduct-0').content,
+  },
+  wallet: {
+    light: () => require('../svgJs/light/wallet-0').content,
+    dark: () => require('../svgJs/dark/wallet-0').content,
+  },
+  walletAsAServiceProduct: {
+    light: () => require('../svgJs/light/walletAsAServiceProduct-0').content,
+    dark: () => require('../svgJs/dark/walletAsAServiceProduct-0').content,
   },
 } as Record<SpotIconName, { light: () => string; dark: () => string }>;