+++ packages/illustrations/src/__generated__/spotIcon/svgJs/dark/privateClientProduct-0.js
@@ -1,0 +1,3 @@
+module.exports = {
+  content: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 32 32"><path fill="#588AF5" d="M31 14H20.998L16 32l15-18Z"/><path fill="#0A0B0D" d="m11 14 5 18L1 14h10Z"/><path fill="#588AF5" d="M1 14h10l5-10H8.207L1 14Z"/><path fill="#464B55" d="M31 14H21L16 4h7.793L31 14Z"/><path fill="#FFFFFF" d="m11 14 5-10 5 10H11Z"/><path fill="#464B55" d="M21 14H11l5 18 5-18Z"/></svg>`,
+};