+++ packages/illustrations/src/__generated__/spotIcon/svgJs/dark/productCoinbaseCard-1.js
@@ -1,0 +1,3 @@
+module.exports = {
+  content: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 32 32"><path fill="#588AF5" d="M22.28 30V2H3.24v28h19.04Z"/><path fill="#0A0B0D" d="M14.61 10.54V5.67h-3.75v4.87h3.75Z"/><path fill="#464B55" d="M15.56 20.19a6.72 6.72 0 1 0 13.44 0 6.72 6.72 0 0 0-13.44 0Z"/><path fill="#0A0B0D" d="M22.28 26.91c-3.71 0-6.72-3.01-6.72-6.72s3.01-6.72 6.72-6.72"/></svg>`,
+};