+++ packages/illustrations/src/__generated__/spotIcon/svgJs/light/borrowProduct-0.js
@@ -1,0 +1,3 @@
+module.exports = {
+  content: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 32 32"><path fill="#00D17F" d="M26 2H6v28h20V2Z"/><path fill="#0052FF" d="M30 8H2v16h28V8Z"/><path fill="#FFFFFF" d="M17 12h-2v8h2v-8Z"/><path fill="#FFFFFF" d="M20 15h-8v2h8v-2Z"/><path fill="#0A0B0D" d="M26 24a4 4 0 0 1 4-4v4h-4ZM6 24a4 4 0 0 0-4-4v4h4ZM26 8a4 4 0 0 0 4 4V8h-4ZM6 8a4 4 0 0 1-4 4V8h4Z"/></svg>`,
+};