+++ packages/illustrations/src/__generated__/spotSquare/data/svgJsMap.ts
@@ -14,20 +14,20 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/accessToAdvancedCharts-2').content,
   },
   addCard: {
-    light: () => require('../svgJs/light/addCard-2').content,
-    dark: () => require('../svgJs/dark/addCard-2').content,
+    light: () => require('../svgJs/light/addCard-3').content,
+    dark: () => require('../svgJs/dark/addCard-3').content,
   },
   addEth: {
     light: () => require('../svgJs/light/addEth-3').content,
     dark: () => require('../svgJs/dark/addEth-3').content,
   },
   addMultipleCrypto: {
-    light: () => require('../svgJs/light/addMultipleCrypto-4').content,
-    dark: () => require('../svgJs/dark/addMultipleCrypto-4').content,
+    light: () => require('../svgJs/light/addMultipleCrypto-5').content,
+    dark: () => require('../svgJs/dark/addMultipleCrypto-5').content,
   },
   addPasswordProtection: {
-    light: () => require('../svgJs/light/addPasswordProtection-3').content,
-    dark: () => require('../svgJs/dark/addPasswordProtection-3').content,
+    light: () => require('../svgJs/light/addPasswordProtection-4').content,
+    dark: () => require('../svgJs/dark/addPasswordProtection-4').content,
   },
   addPhoneNumber: {
     light: () => require('../svgJs/light/addPhoneNumber-2').content,
@@ -46,20 +46,20 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/advancedTradingUi-2').content,
   },
   announcementAdvancedTrading: {
-    light: () => require('../svgJs/light/announcementAdvancedTrading-3').content,
-    dark: () => require('../svgJs/dark/announcementAdvancedTrading-3').content,
+    light: () => require('../svgJs/light/announcementAdvancedTrading-4').content,
+    dark: () => require('../svgJs/dark/announcementAdvancedTrading-4').content,
   },
   appTrackingTransparency: {
     light: () => require('../svgJs/light/appTrackingTransparency-2').content,
     dark: () => require('../svgJs/dark/appTrackingTransparency-2').content,
   },
   assetForward: {
-    light: () => require('../svgJs/light/assetForward-3').content,
-    dark: () => require('../svgJs/dark/assetForward-3').content,
+    light: () => require('../svgJs/light/assetForward-4').content,
+    dark: () => require('../svgJs/dark/assetForward-4').content,
   },
   assetRefresh: {
-    light: () => require('../svgJs/light/assetRefresh-3').content,
-    dark: () => require('../svgJs/dark/assetRefresh-3').content,
+    light: () => require('../svgJs/light/assetRefresh-4').content,
+    dark: () => require('../svgJs/dark/assetRefresh-4').content,
   },
   automaticPayments: {
     light: () => require('../svgJs/light/automaticPayments-2').content,
@@ -86,8 +86,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/boostedCard-4').content,
   },
   borrowLimitsAddressed: {
-    light: () => require('../svgJs/light/borrowLimitsAddressed-4').content,
-    dark: () => require('../svgJs/dark/borrowLimitsAddressed-4').content,
+    light: () => require('../svgJs/light/borrowLimitsAddressed-5').content,
+    dark: () => require('../svgJs/dark/borrowLimitsAddressed-5').content,
   },
   borrowWallet: {
     light: () => require('../svgJs/light/borrowWallet-3').content,
@@ -102,28 +102,28 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/browserExtension-2').content,
   },
   bullishCase: {
-    light: () => require('../svgJs/light/bullishCase-3').content,
-    dark: () => require('../svgJs/dark/bullishCase-3').content,
+    light: () => require('../svgJs/light/bullishCase-4').content,
+    dark: () => require('../svgJs/dark/bullishCase-4').content,
   },
   cardAnnouncement: {
-    light: () => require('../svgJs/light/cardAnnouncement-3').content,
-    dark: () => require('../svgJs/dark/cardAnnouncement-3').content,
+    light: () => require('../svgJs/light/cardAnnouncement-4').content,
+    dark: () => require('../svgJs/dark/cardAnnouncement-4').content,
   },
   cardAutoReload: {
-    light: () => require('../svgJs/light/cardAutoReload-3').content,
-    dark: () => require('../svgJs/dark/cardAutoReload-3').content,
+    light: () => require('../svgJs/light/cardAutoReload-4').content,
+    dark: () => require('../svgJs/dark/cardAutoReload-4').content,
   },
   cardBlocked: {
-    light: () => require('../svgJs/light/cardBlocked-3').content,
-    dark: () => require('../svgJs/dark/cardBlocked-3').content,
+    light: () => require('../svgJs/light/cardBlocked-4').content,
+    dark: () => require('../svgJs/dark/cardBlocked-4').content,
   },
   cardDeclined: {
-    light: () => require('../svgJs/light/cardDeclined-2').content,
-    dark: () => require('../svgJs/dark/cardDeclined-2').content,
+    light: () => require('../svgJs/light/cardDeclined-3').content,
+    dark: () => require('../svgJs/dark/cardDeclined-3').content,
   },
   cardShipped: {
-    light: () => require('../svgJs/light/cardShipped-3').content,
-    dark: () => require('../svgJs/dark/cardShipped-3').content,
+    light: () => require('../svgJs/light/cardShipped-4').content,
+    dark: () => require('../svgJs/dark/cardShipped-4').content,
   },
   cbEthWrappingUnavailable: {
     light: () => require('../svgJs/light/cbEthWrappingUnavailable-2').content,
@@ -138,8 +138,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/coinbaseCardPocket-2').content,
   },
   coinbaseCardSparkle: {
-    light: () => require('../svgJs/light/coinbaseCardSparkle-3').content,
-    dark: () => require('../svgJs/dark/coinbaseCardSparkle-3').content,
+    light: () => require('../svgJs/light/coinbaseCardSparkle-4').content,
+    dark: () => require('../svgJs/dark/coinbaseCardSparkle-4').content,
   },
   coinbaseFees: {
     light: () => require('../svgJs/light/coinbaseFees-2').content,
@@ -198,20 +198,20 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/completeAQuiz-3').content,
   },
   confirmAddress: {
-    light: () => require('../svgJs/light/confirmAddress-3').content,
-    dark: () => require('../svgJs/dark/confirmAddress-3').content,
+    light: () => require('../svgJs/light/confirmAddress-4').content,
+    dark: () => require('../svgJs/dark/confirmAddress-4').content,
   },
   confirmEmail: {
-    light: () => require('../svgJs/light/confirmEmail-3').content,
-    dark: () => require('../svgJs/dark/confirmEmail-3').content,
+    light: () => require('../svgJs/light/confirmEmail-4').content,
+    dark: () => require('../svgJs/dark/confirmEmail-4').content,
   },
   confirmIDCard: {
-    light: () => require('../svgJs/light/confirmIDCard-3').content,
-    dark: () => require('../svgJs/dark/confirmIDCard-3').content,
+    light: () => require('../svgJs/light/confirmIDCard-4').content,
+    dark: () => require('../svgJs/dark/confirmIDCard-4').content,
   },
   confirmSocialSecurity: {
-    light: () => require('../svgJs/light/confirmSocialSecurity-3').content,
-    dark: () => require('../svgJs/dark/confirmSocialSecurity-3').content,
+    light: () => require('../svgJs/light/confirmSocialSecurity-4').content,
+    dark: () => require('../svgJs/dark/confirmSocialSecurity-4').content,
   },
   congratulationsOnEarningCrypto: {
     light: () => require('../svgJs/light/congratulationsOnEarningCrypto-4').content,
@@ -230,8 +230,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/cryptoAndMore-2').content,
   },
   cryptoApps: {
-    light: () => require('../svgJs/light/cryptoApps-5').content,
-    dark: () => require('../svgJs/dark/cryptoApps-5').content,
+    light: () => require('../svgJs/light/cryptoApps-6').content,
+    dark: () => require('../svgJs/dark/cryptoApps-6').content,
   },
   cryptoAssets: {
     light: () => require('../svgJs/light/cryptoAssets-4').content,
@@ -258,8 +258,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/dappWallet-3').content,
   },
   darkModeIntroduction: {
-    light: () => require('../svgJs/light/darkModeIntroduction-3').content,
-    dark: () => require('../svgJs/dark/darkModeIntroduction-3').content,
+    light: () => require('../svgJs/light/darkModeIntroduction-4').content,
+    dark: () => require('../svgJs/dark/darkModeIntroduction-4').content,
   },
   decentralization: {
     light: () => require('../svgJs/light/decentralization-3').content,
@@ -282,16 +282,16 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/defiEarn-2').content,
   },
   defiEarnAnnouncement: {
-    light: () => require('../svgJs/light/defiEarnAnnouncement-4').content,
-    dark: () => require('../svgJs/dark/defiEarnAnnouncement-4').content,
+    light: () => require('../svgJs/light/defiEarnAnnouncement-5').content,
+    dark: () => require('../svgJs/dark/defiEarnAnnouncement-5').content,
   },
   defiHow: {
     light: () => require('../svgJs/light/defiHow-2').content,
     dark: () => require('../svgJs/dark/defiHow-2').content,
   },
   defiNfts: {
-    light: () => require('../svgJs/light/defiNfts-3').content,
-    dark: () => require('../svgJs/dark/defiNfts-3').content,
+    light: () => require('../svgJs/light/defiNfts-4').content,
+    dark: () => require('../svgJs/dark/defiNfts-4').content,
   },
   defiRisk: {
     light: () => require('../svgJs/light/defiRisk-2').content,
@@ -306,8 +306,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/digitalCollectibles-3').content,
   },
   directDepositExcitement: {
-    light: () => require('../svgJs/light/directDepositExcitement-3').content,
-    dark: () => require('../svgJs/dark/directDepositExcitement-3').content,
+    light: () => require('../svgJs/light/directDepositExcitement-4').content,
+    dark: () => require('../svgJs/dark/directDepositExcitement-4').content,
   },
   documentCertified: {
     light: () => require('../svgJs/light/documentCertified-2').content,
@@ -342,8 +342,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/estimatedAmount-2').content,
   },
   eth2SendSell: {
-    light: () => require('../svgJs/light/eth2SendSell-2').content,
-    dark: () => require('../svgJs/dark/eth2SendSell-2').content,
+    light: () => require('../svgJs/light/eth2SendSell-3').content,
+    dark: () => require('../svgJs/dark/eth2SendSell-3').content,
   },
   ethStakeOrWrap: {
     light: () => require('../svgJs/light/ethStakeOrWrap-2').content,
@@ -358,16 +358,16 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/ethStaking-3').content,
   },
   ethStakingRewards: {
-    light: () => require('../svgJs/light/ethStakingRewards-2').content,
-    dark: () => require('../svgJs/dark/ethStakingRewards-2').content,
+    light: () => require('../svgJs/light/ethStakingRewards-3').content,
+    dark: () => require('../svgJs/dark/ethStakingRewards-3').content,
   },
   fileYourCryptoTaxesCheckOther: {
-    light: () => require('../svgJs/light/fileYourCryptoTaxesCheckOther-3').content,
-    dark: () => require('../svgJs/dark/fileYourCryptoTaxesCheckOther-3').content,
+    light: () => require('../svgJs/light/fileYourCryptoTaxesCheckOther-4').content,
+    dark: () => require('../svgJs/dark/fileYourCryptoTaxesCheckOther-4').content,
   },
   fileYourCryptoTaxesOther: {
-    light: () => require('../svgJs/light/fileYourCryptoTaxesOther-3').content,
-    dark: () => require('../svgJs/dark/fileYourCryptoTaxesOther-3').content,
+    light: () => require('../svgJs/light/fileYourCryptoTaxesOther-4').content,
+    dark: () => require('../svgJs/dark/fileYourCryptoTaxesOther-4').content,
   },
   focusLimitOrders: {
     light: () => require('../svgJs/light/focusLimitOrders-2').content,
@@ -398,36 +398,36 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/getStartedInMinutes-4').content,
   },
   giftBoxCrypto: {
-    light: () => require('../svgJs/light/giftBoxCrypto-4').content,
-    dark: () => require('../svgJs/dark/giftBoxCrypto-4').content,
+    light: () => require('../svgJs/light/giftBoxCrypto-5').content,
+    dark: () => require('../svgJs/dark/giftBoxCrypto-5').content,
   },
   gifting: {
-    light: () => require('../svgJs/light/gifting-4').content,
-    dark: () => require('../svgJs/dark/gifting-4').content,
+    light: () => require('../svgJs/light/gifting-5').content,
+    dark: () => require('../svgJs/dark/gifting-5').content,
   },
   globalTransactions: {
     light: () => require('../svgJs/light/globalTransactions-4').content,
     dark: () => require('../svgJs/dark/globalTransactions-4').content,
   },
   guideBullCase: {
-    light: () => require('../svgJs/light/guideBullCase-2').content,
-    dark: () => require('../svgJs/dark/guideBullCase-2').content,
+    light: () => require('../svgJs/light/guideBullCase-3').content,
+    dark: () => require('../svgJs/dark/guideBullCase-3').content,
   },
   guideCryptoBeginner: {
-    light: () => require('../svgJs/light/guideCryptoBeginner-3').content,
-    dark: () => require('../svgJs/dark/guideCryptoBeginner-3').content,
+    light: () => require('../svgJs/light/guideCryptoBeginner-4').content,
+    dark: () => require('../svgJs/dark/guideCryptoBeginner-4').content,
   },
   guideFiveThings: {
-    light: () => require('../svgJs/light/guideFiveThings-3').content,
-    dark: () => require('../svgJs/dark/guideFiveThings-3').content,
+    light: () => require('../svgJs/light/guideFiveThings-4').content,
+    dark: () => require('../svgJs/dark/guideFiveThings-4').content,
   },
   guideNftDefi: {
-    light: () => require('../svgJs/light/guideNftDefi-3').content,
-    dark: () => require('../svgJs/dark/guideNftDefi-3').content,
+    light: () => require('../svgJs/light/guideNftDefi-4').content,
+    dark: () => require('../svgJs/dark/guideNftDefi-4').content,
   },
   guideStartInvesting: {
-    light: () => require('../svgJs/light/guideStartInvesting-3').content,
-    dark: () => require('../svgJs/dark/guideStartInvesting-3').content,
+    light: () => require('../svgJs/light/guideStartInvesting-4').content,
+    dark: () => require('../svgJs/dark/guideStartInvesting-4').content,
   },
   hardwareWallets: {
     light: () => require('../svgJs/light/hardwareWallets-2').content,
@@ -446,8 +446,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/insuranceProtection-3').content,
   },
   interestForYou: {
-    light: () => require('../svgJs/light/interestForYou-4').content,
-    dark: () => require('../svgJs/dark/interestForYou-4').content,
+    light: () => require('../svgJs/light/interestForYou-5').content,
+    dark: () => require('../svgJs/dark/interestForYou-5').content,
   },
   invest: {
     light: () => require('../svgJs/light/invest-3').content,
@@ -466,8 +466,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/linkingYourWalletToYourCoinbaseAccount-3').content,
   },
   miniGift: {
-    light: () => require('../svgJs/light/miniGift-3').content,
-    dark: () => require('../svgJs/dark/miniGift-3').content,
+    light: () => require('../svgJs/light/miniGift-4').content,
+    dark: () => require('../svgJs/dark/miniGift-4').content,
   },
   mining: {
     light: () => require('../svgJs/light/mining-2').content,
@@ -478,8 +478,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/moneyDecentralized-3').content,
   },
   moneyRewards: {
-    light: () => require('../svgJs/light/moneyRewards-4').content,
-    dark: () => require('../svgJs/dark/moneyRewards-4').content,
+    light: () => require('../svgJs/light/moneyRewards-5').content,
+    dark: () => require('../svgJs/dark/moneyRewards-5').content,
   },
   multicoinSupport: {
     light: () => require('../svgJs/light/multicoinSupport-4').content,
@@ -494,12 +494,12 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/multipleAccountsWalletsForOneUser-3').content,
   },
   nft: {
-    light: () => require('../svgJs/light/nft-4').content,
-    dark: () => require('../svgJs/dark/nft-4').content,
+    light: () => require('../svgJs/light/nft-5').content,
+    dark: () => require('../svgJs/dark/nft-5').content,
   },
   nftTag: {
-    light: () => require('../svgJs/light/nftTag-3').content,
-    dark: () => require('../svgJs/dark/nftTag-3').content,
+    light: () => require('../svgJs/light/nftTag-4').content,
+    dark: () => require('../svgJs/dark/nftTag-4').content,
   },
   noFees: {
     light: () => require('../svgJs/light/noFees-3').content,
@@ -514,24 +514,24 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/notificationsAlt-2').content,
   },
   nuxChecklist: {
-    light: () => require('../svgJs/light/nuxChecklist-4').content,
-    dark: () => require('../svgJs/dark/nuxChecklist-4').content,
+    light: () => require('../svgJs/light/nuxChecklist-5').content,
+    dark: () => require('../svgJs/dark/nuxChecklist-5').content,
   },
   nuxEarnCrypto: {
-    light: () => require('../svgJs/light/nuxEarnCrypto-3').content,
-    dark: () => require('../svgJs/dark/nuxEarnCrypto-3').content,
+    light: () => require('../svgJs/light/nuxEarnCrypto-4').content,
+    dark: () => require('../svgJs/dark/nuxEarnCrypto-4').content,
   },
   nuxEarnYield: {
-    light: () => require('../svgJs/light/nuxEarnYield-4').content,
-    dark: () => require('../svgJs/dark/nuxEarnYield-4').content,
+    light: () => require('../svgJs/light/nuxEarnYield-5').content,
+    dark: () => require('../svgJs/dark/nuxEarnYield-5').content,
   },
   nuxPopularAssets: {
-    light: () => require('../svgJs/light/nuxPopularAssets-3').content,
-    dark: () => require('../svgJs/dark/nuxPopularAssets-3').content,
+    light: () => require('../svgJs/light/nuxPopularAssets-4').content,
+    dark: () => require('../svgJs/dark/nuxPopularAssets-4').content,
   },
   nuxRecurringBuys: {
-    light: () => require('../svgJs/light/nuxRecurringBuys-3').content,
-    dark: () => require('../svgJs/dark/nuxRecurringBuys-3').content,
+    light: () => require('../svgJs/light/nuxRecurringBuys-4').content,
+    dark: () => require('../svgJs/dark/nuxRecurringBuys-4').content,
   },
   offersEmpty: {
     light: () => require('../svgJs/light/offersEmpty-4').content,
@@ -554,12 +554,12 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/p2pPayments-3').content,
   },
   performance: {
-    light: () => require('../svgJs/light/performance-3').content,
-    dark: () => require('../svgJs/dark/performance-3').content,
+    light: () => require('../svgJs/light/performance-4').content,
+    dark: () => require('../svgJs/dark/performance-4').content,
   },
   phoneNotifications: {
-    light: () => require('../svgJs/light/phoneNotifications-4').content,
-    dark: () => require('../svgJs/dark/phoneNotifications-4').content,
+    light: () => require('../svgJs/light/phoneNotifications-5').content,
+    dark: () => require('../svgJs/dark/phoneNotifications-5').content,
   },
   phoneNumber: {
     light: () => require('../svgJs/light/phoneNumber-1').content,
@@ -582,20 +582,20 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/poweredByEthereum-4').content,
   },
   priceAlerts: {
-    light: () => require('../svgJs/light/priceAlerts-3').content,
-    dark: () => require('../svgJs/dark/priceAlerts-3').content,
+    light: () => require('../svgJs/light/priceAlerts-4').content,
+    dark: () => require('../svgJs/dark/priceAlerts-4').content,
   },
   primeDeFi: {
-    light: () => require('../svgJs/light/primeDeFi-3').content,
-    dark: () => require('../svgJs/dark/primeDeFi-3').content,
+    light: () => require('../svgJs/light/primeDeFi-4').content,
+    dark: () => require('../svgJs/dark/primeDeFi-4').content,
   },
   primeEarn: {
-    light: () => require('../svgJs/light/primeEarn-3').content,
-    dark: () => require('../svgJs/dark/primeEarn-3').content,
+    light: () => require('../svgJs/light/primeEarn-4').content,
+    dark: () => require('../svgJs/dark/primeEarn-4').content,
   },
   primeStaking: {
-    light: () => require('../svgJs/light/primeStaking-3').content,
-    dark: () => require('../svgJs/dark/primeStaking-3').content,
+    light: () => require('../svgJs/light/primeStaking-4').content,
+    dark: () => require('../svgJs/dark/primeStaking-4').content,
   },
   quickAndSimple: {
     light: () => require('../svgJs/light/quickAndSimple-4').content,
@@ -606,28 +606,28 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/readyToTrade-2').content,
   },
   recommendInvestments: {
-    light: () => require('../svgJs/light/recommendInvestments-3').content,
-    dark: () => require('../svgJs/dark/recommendInvestments-3').content,
+    light: () => require('../svgJs/light/recommendInvestments-4').content,
+    dark: () => require('../svgJs/dark/recommendInvestments-4').content,
   },
   referralsPeople: {
     light: () => require('../svgJs/light/referralsPeople-3').content,
     dark: () => require('../svgJs/dark/referralsPeople-3').content,
   },
   refreshMobileApp: {
-    light: () => require('../svgJs/light/refreshMobileApp-3').content,
-    dark: () => require('../svgJs/dark/refreshMobileApp-3').content,
+    light: () => require('../svgJs/light/refreshMobileApp-4').content,
+    dark: () => require('../svgJs/dark/refreshMobileApp-4').content,
   },
   retailUSDCRewards: {
     light: () => require('../svgJs/light/retailUSDCRewards-3').content,
     dark: () => require('../svgJs/dark/retailUSDCRewards-3').content,
   },
   rewardExpiring: {
-    light: () => require('../svgJs/light/rewardExpiring-3').content,
-    dark: () => require('../svgJs/dark/rewardExpiring-3').content,
+    light: () => require('../svgJs/light/rewardExpiring-4').content,
+    dark: () => require('../svgJs/dark/rewardExpiring-4').content,
   },
   saveTheDate: {
-    light: () => require('../svgJs/light/saveTheDate-3').content,
-    dark: () => require('../svgJs/dark/saveTheDate-3').content,
+    light: () => require('../svgJs/light/saveTheDate-4').content,
+    dark: () => require('../svgJs/dark/saveTheDate-4').content,
   },
   secureAndTrusted: {
     light: () => require('../svgJs/light/secureAndTrusted-3').content,
@@ -654,8 +654,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/semiCustodial-4').content,
   },
   sendCryptoFaster: {
-    light: () => require('../svgJs/light/sendCryptoFaster-3').content,
-    dark: () => require('../svgJs/dark/sendCryptoFaster-3').content,
+    light: () => require('../svgJs/light/sendCryptoFaster-4').content,
+    dark: () => require('../svgJs/dark/sendCryptoFaster-4').content,
   },
   shareOnSocialMedia: {
     light: () => require('../svgJs/light/shareOnSocialMedia-3').content,
@@ -666,8 +666,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/sidechain-2').content,
   },
   sparkleToken: {
-    light: () => require('../svgJs/light/sparkleToken-4').content,
-    dark: () => require('../svgJs/dark/sparkleToken-4').content,
+    light: () => require('../svgJs/light/sparkleToken-5').content,
+    dark: () => require('../svgJs/dark/sparkleToken-5').content,
   },
   stableValue: {
     light: () => require('../svgJs/light/stableValue-3').content,
@@ -678,8 +678,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/staking-3').content,
   },
   starToken: {
-    light: () => require('../svgJs/light/starToken-3').content,
-    dark: () => require('../svgJs/dark/starToken-3').content,
+    light: () => require('../svgJs/light/starToken-4').content,
+    dark: () => require('../svgJs/dark/starToken-4').content,
   },
   startToday: {
     light: () => require('../svgJs/light/startToday-3').content,
@@ -702,12 +702,12 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/switchAdvancedToSimpleTrading-2').content,
   },
   switchReward: {
-    light: () => require('../svgJs/light/switchReward-3').content,
-    dark: () => require('../svgJs/dark/switchReward-3').content,
+    light: () => require('../svgJs/light/switchReward-4').content,
+    dark: () => require('../svgJs/dark/switchReward-4').content,
   },
   taxDocuments: {
-    light: () => require('../svgJs/light/taxDocuments-3').content,
-    dark: () => require('../svgJs/dark/taxDocuments-3').content,
+    light: () => require('../svgJs/light/taxDocuments-4').content,
+    dark: () => require('../svgJs/dark/taxDocuments-4').content,
   },
   taxesDetails: {
     light: () => require('../svgJs/light/taxesDetails-3').content,
@@ -718,8 +718,8 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/tradeImmediately-2').content,
   },
   transferringCrypto: {
-    light: () => require('../svgJs/light/transferringCrypto-4').content,
-    dark: () => require('../svgJs/dark/transferringCrypto-4').content,
+    light: () => require('../svgJs/light/transferringCrypto-5').content,
+    dark: () => require('../svgJs/dark/transferringCrypto-5').content,
   },
   trendingHotAssets: {
     light: () => require('../svgJs/light/trendingHotAssets-2').content,
@@ -738,12 +738,12 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/verifyInfo-2').content,
   },
   waitlistSignup: {
-    light: () => require('../svgJs/light/waitlistSignup-3').content,
-    dark: () => require('../svgJs/dark/waitlistSignup-3').content,
+    light: () => require('../svgJs/light/waitlistSignup-4').content,
+    dark: () => require('../svgJs/dark/waitlistSignup-4').content,
   },
   walletApp: {
-    light: () => require('../svgJs/light/walletApp-4').content,
-    dark: () => require('../svgJs/dark/walletApp-4').content,
+    light: () => require('../svgJs/light/walletApp-5').content,
+    dark: () => require('../svgJs/dark/walletApp-5').content,
   },
   walletNotifications: {
     light: () => require('../svgJs/light/walletNotifications-2').content,
@@ -762,8 +762,12 @@ const svgJsMap = {
     dark: () => require('../svgJs/dark/wrapEth-3').content,
   },
   yieldCenter: {
-    light: () => require('../svgJs/light/yieldCenter-4').content,
-    dark: () => require('../svgJs/dark/yieldCenter-4').content,
+    light: () => require('../svgJs/light/yieldCenter-5').content,
+    dark: () => require('../svgJs/dark/yieldCenter-5').content,
+  },
+  yieldCenterUSDC: {
+    light: () => require('../svgJs/light/yieldCenterUSDC-0').content,
+    dark: () => require('../svgJs/dark/yieldCenterUSDC-0').content,
   },
 } as Record<SpotSquareName, { light: () => string; dark: () => string }>;