+++ packages/web/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file!
 
 <!-- template-start -->
 
+## Unreleased
+
+#### 📘 Misc
+
+- Regenerate icons and illustrations stories.
+
 ## 5.16.5 ((10/26/2023, 07:13 PM PST))
 
 This is an artificial version bump with no new change.